requirements.txt
singer-python>=5.0.12
google-cloud-pubsub>=1.1.0
pytest>=5.3.2
